% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid.R
\name{bid}
\alias{bid}
\title{Bid Function}
\usage{
bid(x_i, n, c, k, mu_0, sigma_0)
}
\arguments{
\item{x_i}{The bidder's signal.}

\item{n}{The number of bidders.}

\item{c}{The cost of default.}

\item{k}{The coefficient on the signal standard variance.}

\item{mu_0}{The prior mean.}

\item{sigma_0}{The prior standard deviation.}
}
\value{
The equilibrium bid \eqn{\beta(x_i)}
}
\description{
This function will calculate the optimal bid for a signal in an auction by finding the bid where the \link{eqm} function takes value zero.
}
