% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid_mats.R
\name{bid_mats}
\alias{bid_mats}
\title{Calculates matrices of optimal bids for interpolation.}
\usage{
bid_mats(x_seq, c_seq, n_seq, k, mu_0, sigma_0)
}
\arguments{
\item{x_seq}{Sequence of signals for which to compute bids for interpolation.}

\item{c_seq}{Sequence of default costs for which to compute bids for interpolation.}

\item{n_seq}{The possible numbers of bidders.}

\item{k}{The coefficient on the signal standard variance.}

\item{mu_0}{The prior mean.}

\item{sigma_0}{The prior SD.}
}
\value{
A list of bidding matrices for each set of parameters.
}
\description{
Calculates matrices of optimal bids for interpolation.
}
