% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid_vec.R
\name{bid_vec}
\alias{bid_vec}
\title{Generate The Optimal Probability-Weighted Bids}
\usage{
bid_vec(
  mu_t,
  sigma_t,
  x_seq,
  c_seq,
  c,
  lambda_t,
  n_max,
  k,
  mat,
  mu_base = 0,
  sigma_base = 1
)
}
\arguments{
\item{mu_t}{The auction's value of the mean of the common value distribution.}

\item{sigma_t}{The auction's value of the standard deviation of the common value distribution.}

\item{x_seq}{The grid of signals.}

\item{c_seq}{The grid of default costs.}

\item{c}{The auction's nominal cost of default}

\item{lambda_t}{The auction's value of the poisson parameter.}

\item{n_max}{The maximum number of bidders in the auction.}

\item{k}{The coefficient on the signal standard deviation}

\item{mat}{The interpolation matrices.}

\item{mu_base}{The mean of the base case auction.}

\item{sigma_base}{The SD of the base case auction.}
}
\value{
Returns a vector of normalized bids for the auction's normalized cost of default.
}
\description{
This function expects that all arguments are adjusted to the units of the bid matrix.
}
