% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bid_prep_cost.R
\name{calc_bid_prep_cost}
\alias{calc_bid_prep_cost}
\title{Calculate the bid preparation cost for an auction}
\usage{
calc_bid_prep_cost(
  dt,
  mu_t,
  sigma_t,
  lambda_t,
  r_t,
  k,
  c,
  interps,
  x_seq,
  n_seq,
  f_weights
)
}
\arguments{
\item{dt}{Simulated dataframe}

\item{mu_t}{Auction's mean}

\item{sigma_t}{Auction's SD}

\item{lambda_t}{Auction's lambda parameter}

\item{r_t}{Auction's reserve}

\item{k}{Signal standard deviation coefficient}

\item{c}{Cost of default}

\item{interps}{The interpolations for bidding}

\item{x_seq}{The sequence of signals}

\item{n_seq}{The possible number of bidders}

\item{f_weights}{The probability of being the high bidder}
}
\value{
The bid preparation cost for an auction
}
\description{
Calculate the bid preparation cost for an auction
}
