% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_f.R
\name{calc_f}
\alias{calc_f}
\title{Helper Function for "Degrees of Freedom" for Number of Bidders}
\usage{
calc_f(n)
}
\arguments{
\item{n}{Number of bidders.}
}
\value{
The number of "free" bidders.
}
\description{
If \eqn{n>2}, we want to fix the first and second-order statistics. If \eqn{n=2}, we want to fix only the first-order statistic.
}
