% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_phi_grid.R
\name{calc_phi_grid}
\alias{calc_phi_grid}
\title{Calculate the Inverse Bidding Function}
\usage{
calc_phi_grid(
  mu_t,
  sigma_t,
  mu_0,
  sigma_0,
  x_seq,
  c_seq,
  c_norm,
  lambda_t,
  n_max,
  k,
  mat
)
}
\arguments{
\item{mu_t}{The auction's value of the mean of the common value distribution.}

\item{sigma_t}{The auction's value of the standard deviation of the common value distribution.}

\item{mu_0}{The prior value of the value mean.}

\item{sigma_0}{The prior value of value SD.}

\item{x_seq}{The grid of signals.}

\item{c_seq}{The grid of default costs.}

\item{c_norm}{The auction's cost of default, normalized.}

\item{lambda_t}{The auction's value of the poisson parameter.}

\item{n_max}{The maximum number of bidders in the auction.}

\item{k}{The coefficient on the signal standard deviation}

\item{mat}{The interpolation matrices.}
}
\value{
An interpolation object for calculating the inverse of the bidding function.
}
\description{
Calculate the Inverse Bidding Function
}
