% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_phi_individual.R
\name{calc_phi_individual}
\alias{calc_phi_individual}
\title{Calculate inverse bidding function}
\usage{
calc_phi_individual(
  mu_t,
  sigma_t,
  lambda_t,
  k,
  c,
  interps,
  x_seq,
  n_seq,
  f_weights
)
}
\arguments{
\item{mu_t}{a}

\item{sigma_t}{a}

\item{lambda_t}{a}

\item{k}{a}

\item{c}{a}

\item{interps}{a}

\item{x_seq}{a}

\item{n_seq}{a}

\item{f_weights}{a}
}
\value{
spline
}
\description{
Calculate inverse bidding function
}
