% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual_no_resale.R
\name{counterfactual_no_resale}
\alias{counterfactual_no_resale}
\title{Compute the counterfactual for an auction under the assumption of no resale}
\usage{
counterfactual_no_resale(
  c,
  dt,
  bid_prep_cost,
  mu_t,
  sigma_t,
  lambda_t,
  r_t,
  k,
  interps,
  x_seq,
  n_seq,
  f_weights
)
}
\arguments{
\item{c}{The default cost in the counterfactual}

\item{dt}{The datatable of simulations}

\item{bid_prep_cost}{The bid preparation cost for the auction}

\item{mu_t}{The auction's mean}

\item{sigma_t}{The auction's standard deviation}

\item{lambda_t}{The auction's value of lambda}

\item{r_t}{The auction's reserve price}

\item{k}{The coefficient on signal standard deviation}

\item{interps}{The interpolation grids}

\item{x_seq}{The sequence of signals}

\item{n_seq}{The possible number of bidders}

\item{f_weights}{The probability of being the high bidder}
}
\value{
A list of revenue to the seller and the equilibrium lambda for the auction
}
\description{
Compute the counterfactual for an auction under the assumption of no resale
}
