% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_auction_params.R
\name{create_auction_params}
\alias{create_auction_params}
\title{Add the implied auction parameters to the dataset}
\usage{
create_auction_params(
  auction_data,
  L_INT,
  L_N_IMAGE,
  L_YEAR,
  MU_INT,
  MU_N_IMAGE,
  MU_YEAR,
  SIGMA_INT,
  SIGMA_N_IMAGE,
  SIGMA_YEAR
)
}
\arguments{
\item{auction_data}{The item characteristics}

\item{L_INT}{The intercept for lambda}

\item{L_N_IMAGE}{The coefficient on images for lambda}

\item{L_YEAR}{The coefficient on year for lambda}

\item{MU_INT}{The intercept for mu}

\item{MU_N_IMAGE}{The coefficient on images for lambda}

\item{MU_YEAR}{The coefficient on year for lambda}

\item{SIGMA_INT}{The intercept for sigma}

\item{SIGMA_N_IMAGE}{The coefficient for images for sigma}

\item{SIGMA_YEAR}{The coefficient for year for sigma}
}
\value{
A dataframe with all the auction parameters
}
\description{
Add the implied auction parameters to the dataset
}
