% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ev_splines.R
\name{create_ev_splines}
\alias{create_ev_splines}
\title{Calculate the expected value splines for the cutoff signal}
\usage{
create_ev_splines(k, n_max)
}
\arguments{
\item{k}{The multiplier on the signal SD}

\item{n_max}{The maximum number of bidders}
}
\value{
The cutoff signal x*
}
\description{
Calculate the expected value splines for the cutoff signal
}
