% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_interp_mats.R
\name{create_interp_mats}
\alias{create_interp_mats}
\title{Create the matrices for interpolation}
\usage{
create_interp_mats(k, n_max, x_vals, diff_vals)
}
\arguments{
\item{k}{The coefficient on the standard deviation}

\item{n_max}{The maximum number of bidders}

\item{x_vals}{The sequence of signals}

\item{diff_vals}{The sequence of values to "look back" from a signal}
}
\value{
A list of interpolation matrices for each number of bidders
}
\description{
Create the matrices for interpolation
}
