% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sim_df.R
\name{create_sim_df}
\alias{create_sim_df}
\title{Creates a dataframe of simulations}
\usage{
create_sim_df(k, n_sims = 100)
}
\arguments{
\item{k}{The coefficient on the signal standard deviation}

\item{n_sims}{The number of simulations}
}
\value{
A datatable of simulations
}
\description{
Creates a dataframe of simulations
}
