% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_data.R
\name{create_test_data}
\alias{create_test_data}
\title{Generate test data for algorithm}
\usage{
create_test_data(n_sims, seed = 42069)
}
\arguments{
\item{n_sims}{The number of observations to generate}

\item{seed}{The random seed. Defaults to 42069.}
}
\value{
A dataframe of simulated auction data
}
\description{
Generate test data for algorithm
}
