% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_prob.R
\name{default_prob}
\alias{default_prob}
\title{The probability of default}
\usage{
default_prob(v, mu_t, sigma_t, lambda_t, bids, phi, x_s, c, k, n_max)
}
\arguments{
\item{v}{The item's value}

\item{mu_t}{The prior mean of the value distribution}

\item{sigma_t}{The prior SD of the value distribution}

\item{lambda_t}{The parameter governing the Poisson entry process}

\item{bids}{Observed bids}

\item{phi}{Inverse bidding function}

\item{x_s}{Cutoff signal}

\item{c}{Cost of default}

\item{k}{Multiplier on signal SD}

\item{n_max}{The maximum number of bidders}
}
\value{
The probability of default
}
\description{
The probability of default
}
