% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/den.R
\name{den}
\alias{den}
\title{The denominator of the expected value expression}
\usage{
den(v, x, n, mu_0, sigma_0, k)
}
\arguments{
\item{v}{The value of the object}

\item{x}{The signal}

\item{n}{The number of bidders}

\item{mu_0}{The prior mean of the value distribution}

\item{sigma_0}{The prior SD of the value distribution}

\item{k}{The multiplier on the SD of the signals}
}
\value{
The probability of being the highest bidder
}
\description{
The denominator of the expected value expression
}
