% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood_grid.R
\name{evaluate_likelihood_grid}
\alias{evaluate_likelihood_grid}
\title{Calculate the likelihood of the data for a set of parameters}
\usage{
evaluate_likelihood_grid(
  auction_data,
  MU_INT,
  MU_N_IMAGE,
  SIGMA_INT,
  SIGMA_N_IMAGE,
  K,
  C,
  L_INT,
  L_N_IMAGE,
  D_R,
  N_MAX,
  par = FALSE
)
}
\arguments{
\item{auction_data}{The dataset to calculate likelihood over}

\item{MU_INT}{The intercept of the \code{\eqn{\mu}} parameter}

\item{MU_N_IMAGE}{The coefficient on \code{n_images} of the \code{\eqn{\mu}} parameter}

\item{SIGMA_INT}{The intercept of the \code{\eqn{\sigma}} parameter}

\item{SIGMA_N_IMAGE}{The coefficient on \code{n_images} of the \code{\eqn{\sigma}} parameter}

\item{K}{The multiplier on}

\item{C}{a}

\item{L_INT}{a}

\item{L_N_IMAGE}{a}

\item{D_R}{a}

\item{N_MAX}{a}

\item{par}{a}
}
\value{
The log-likelihood of observing the data
}
\description{
Calculate the likelihood of the data for a set of parameters
}
