% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f}
\alias{f}
\title{Probability of a sample mean given a signal}
\usage{
f(v, x_bar, x_i, n_signals, k)
}
\arguments{
\item{v}{The value of the item}

\item{x_bar}{The sample mean of the signals}

\item{x_i}{The bidder's signal}

\item{n_signals}{The number of bidders}

\item{k}{The coefficient on the signal standard deviation}
}
\value{
The probability of a sample mean given a signal
}
\description{
Probability of a sample mean given a signal
}
