% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_y_y.R
\name{f_y_y}
\alias{f_y_y}
\title{Calculate the PDF of the Second-Order Statistic}
\usage{
f_y_y(y, mu_0, sigma_0, n, k)
}
\arguments{
\item{y}{The point where we assess the distribution.}

\item{mu_0}{The prior mean.}

\item{sigma_0}{The prior SD.}

\item{n}{The number of bidders.}

\item{k}{The coefficient on the signal standard variance.}
}
\value{
The PDF of the second-order statistic at \code{y}
}
\description{
Calculate the PDF of the Second-Order Statistic
}
