% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moments.R
\name{get_moments}
\alias{get_moments}
\title{Get the Moments of the Conditional Distribution}
\usage{
get_moments(my_signal, n, interp_mats, x_vals, diff_vals)
}
\arguments{
\item{my_signal}{The bidder's signal}

\item{n}{The number of competitors}

\item{interp_mats}{The matrices over which bicubic interpolation is computed}

\item{x_vals}{The sequence of signals for which to compute the moments}

\item{diff_vals}{The sequence of values to "look back" from the bidder's signal}
}
\value{
A list of matrices with the CDF and conditional expectation
}
\description{
Get the Moments of the Conditional Distribution
}
