% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_summary.R
\name{get_sim_summary}
\alias{get_sim_summary}
\title{Summarise the simulations for an auction}
\usage{
get_sim_summary(
  dt,
  mu_t,
  sigma_t,
  lambda_t,
  r_t,
  k,
  c,
  interps,
  x_seq,
  n_seq,
  f_weights
)
}
\arguments{
\item{dt}{Simulated dataframe}

\item{mu_t}{Auction's mean}

\item{sigma_t}{Auction's SD}

\item{lambda_t}{Auction's lambda parameter}

\item{r_t}{Auction's reserve}

\item{k}{Signal standard deviation coefficient}

\item{c}{Cost of default}

\item{interps}{The interpolations for bidding}

\item{x_seq}{The sequence of signals}

\item{n_seq}{The possible number of bidders}

\item{f_weights}{The probability of being the high bidder}
}
\value{
A dataframe of outcomes for each possible number of bidders
}
\description{
Summarise the simulations for an auction
}
