% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.R
\name{h}
\alias{h}
\title{Calculate the posterior mean from a signal and sample mean}
\usage{
h(x_i, x_bar, mu_0, n, k)
}
\arguments{
\item{x_i}{The signal}

\item{x_bar}{The sample mean}

\item{mu_0}{The prior mean}

\item{n}{The number of bidders}

\item{k}{The coefficient on the standard deviation}
}
\value{
The value of the posterior mean
}
\description{
Calculate the posterior mean from a signal and sample mean
}
