% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_inv.R
\name{h_inv}
\alias{h_inv}
\title{Calculate the inverse transform of signals to posterior mean.}
\usage{
h_inv(mu_1, mu_0, x_i, n, k)
}
\arguments{
\item{mu_1}{The posterior mean.}

\item{mu_0}{The prior mean.}

\item{x_i}{The bidder's signal.}

\item{n}{The number of bidders.}

\item{k}{The coefficient on the signal standard variance.}
}
\value{
The sample mean required to generate this value of the posterior, given a bidder's signal.
}
\description{
Given the posterior mean is updated via \deqn{\mu_1 = h(\overline{x}) = \frac{k\mu_0 + n\overline{x}}{n+k}}
We have that \deqn{\overline{x}_{n-1} = h^{-1}(\mu_1) = \frac{(n+k)\mu_1 - k\mu_0 - x_i}{n-1}}
}
