% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc}
\alias{mcmc}
\title{Estimate MCMC}
\usage{
mcmc(data, n_steps, params, n_max, par = FALSE)
}
\arguments{
\item{data}{The auction data}

\item{n_steps}{The number of draws}

\item{params}{The vector of parameters}

\item{n_max}{The maximum number of bidders}

\item{par}{Whether or not to run the estimation in parallel}
}
\value{
A list of draws and parameters that caused errors.
}
\description{
Estimate MCMC
}
