% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{MLE}
\usage{
mle(init_params, auction_data)
}
\arguments{
\item{init_params}{The initial parameter vector}

\item{auction_data}{The data over which to run the optimization}
}
\value{
The optimization object
}
\description{
MLE
}
