% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_obj.R
\name{mle_obj}
\alias{mle_obj}
\title{MLE Objective Function}
\usage{
mle_obj(param_vec, auction_data, par)
}
\arguments{
\item{param_vec}{The vector of parameters}

\item{auction_data}{The data}

\item{par}{Whether the routine should be run in parallel}
}
\value{
The likelihood
}
\description{
MLE Objective Function
}
