% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dist.R
\name{post_dist}
\alias{post_dist}
\title{Posterior Distribution of Item's Value}
\usage{
post_dist(x, mu_0, sigma_0, k, fun = "pdf")
}
\arguments{
\item{x}{The observed data.}

\item{mu_0}{The prior mean.}

\item{sigma_0}{The prior standard deviation.}

\item{k}{The coefficient on the signal standard deviation}

\item{fun}{Whether to return the PDF or CDF}
}
\value{
A \code{\link[stats]{dnorm}} object with the appropriate hyperparameters.
}
\description{
Posterior Distribution of Item's Value
}
