% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_exp.R
\name{post_exp}
\alias{post_exp}
\title{Calculate the Expected Value for a Given Signal}
\usage{
post_exp(x, mu_0, sigma_0, n, k)
}
\arguments{
\item{x}{The signal.}

\item{mu_0}{The prior mean.}

\item{sigma_0}{The prior SD.}

\item{n}{The number of bidders.}

\item{k}{The coefficient on the signal standard variance.}
}
\value{
The expectation of the value given signal \code{x}
}
\description{
Calculate the Expected Value for a Given Signal
}
