% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_mean_cdf.R
\name{post_mean_cdf}
\alias{post_mean_cdf}
\title{Calculate the CDF of the distribution of the posterior mean.}
\usage{
post_mean_cdf(mu_1, mu_0, sigma_0, x_i, n, k)
}
\arguments{
\item{mu_1}{The point on the distribution where the CDF is assessed.}

\item{mu_0}{The prior mean.}

\item{sigma_0}{The prior SD.}

\item{x_i}{The bidder's signal.}

\item{n}{The number of bidders.}

\item{k}{The coefficient on the signal standard variance.}
}
\value{
The CDF of the distribution of the expected posterior mean, evaluated at \code{mu_1}.
}
\description{
Calculate the CDF of the distribution of the posterior mean.
}
