% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_mean_exp.R
\name{post_mean_exp}
\alias{post_mean_exp}
\title{Calculate the expectation of the distribution of the posterior mean.}
\usage{
post_mean_exp(x_i, mu_0, sigma_0, n, k, lb = -Inf)
}
\arguments{
\item{x_i}{The bidder's signal.}

\item{mu_0}{The prior mean.}

\item{sigma_0}{The prior SD.}

\item{n}{The number of bidders.}

\item{k}{The coefficient on the signal standard variance.}

\item{lb}{The lower truncation of the distribution (assumed to not exist if not specified).}
}
\value{
The expectation of the posterior mean, given that it's less than the bidder's signal.
}
\description{
Calculate the expectation of the distribution of the posterior mean.
}
