% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\title{Prior}
\usage{
prior(params)
}
\arguments{
\item{params}{The parameters}
}
\value{
A boolean indicating whether a parameter vector is assigned non-zero probability under the prior
}
\description{
Prior
}
