% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_bids.R
\name{prob_bids}
\alias{prob_bids}
\title{The probability of observing bids}
\usage{
prob_bids(v, j, n, mu_t, sigma_t, lambda_t, bids, default, phi, x_s, k)
}
\arguments{
\item{v}{The value of the item}

\item{j}{The total number of bidders}

\item{n}{The number of observed bidders}

\item{mu_t}{The mean of the prior distribution}

\item{sigma_t}{The standard deviation of the prior distribution}

\item{lambda_t}{The parameter governing the poisson entry process}

\item{bids}{The bids}

\item{default}{Whether or not default occurred}

\item{phi}{The inverse bidding function}

\item{x_s}{The cutoff signal}

\item{k}{The multiplier on the signal standard deviation}
}
\value{
The probability of observing the bids that occurred
}
\description{
The probability of observing bids
}
