% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_default_given_bids.R
\name{prob_default_given_bids}
\alias{prob_default_given_bids}
\title{Calculate the probability of default given bids.}
\usage{
prob_default_given_bids(
  v,
  mu_t,
  sigma_t,
  lambda_t,
  bids,
  phi,
  default,
  x_s,
  c,
  k,
  n_max,
  dr
)
}
\arguments{
\item{v}{The value}

\item{mu_t}{The prior on the mean of the value distribution}

\item{sigma_t}{The prior on the SD of the value distribution}

\item{lambda_t}{The entry probability parameter}

\item{bids}{The bids}

\item{phi}{The inverse bid function}

\item{default}{Whether or not default occurred}

\item{x_s}{The cutoff signal}

\item{c}{The cost of default}

\item{k}{The multiplier on the SD of the bidder's signals}

\item{n_max}{The max number of bidders}

\item{dr}{The base rate default probability}
}
\value{
A default probability for a set of bids
}
\description{
Calculate the probability of default given bids.
}
