% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemann_sums.R
\name{riemann_sums}
\alias{riemann_sums}
\title{Integrate via Riemann Sums}
\usage{
riemann_sums(f, lower, upper, n_bins, ...)
}
\arguments{
\item{f}{The function to be integrated}

\item{lower}{The lower bound of the integral}

\item{upper}{The upper bound of the integral}

\item{n_bins}{The number of bins to use for integration}

\item{...}{Other arguments to be passed to the integrating function}
}
\value{
The area under the function
}
\description{
Integrate via Riemann Sums
}
