% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_default.R
\name{simulated_default}
\alias{simulated_default}
\title{Create default decisions for simulated data}
\usage{
simulated_default(
  signals,
  bids,
  xs_vals,
  lambda,
  mu,
  sigma,
  max_bidders,
  c,
  k,
  r
)
}
\arguments{
\item{signals}{The signals}

\item{bids}{The bids}

\item{xs_vals}{The cutoff signal}

\item{lambda}{The parameter governing the Poisson entry process}

\item{mu}{The mean of the prior distribution}

\item{sigma}{The standard deviation of the prior distribution}

\item{max_bidders}{The maximum number of bidders}

\item{c}{The cost of default}

\item{k}{The multiplier on the signals}

\item{r}{The reserve price}
}
\value{
Whether or not the winning bidder defaults
}
\description{
Create default decisions for simulated data
}
