% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnorm_cdf.R
\name{tnorm_cdf}
\alias{tnorm_cdf}
\title{Calculate the CDF of a Normal Truncated Above}
\usage{
tnorm_cdf(x, mu, sigma, k, ub)
}
\arguments{
\item{x}{The point where we assess the CDF.}

\item{mu}{The distribution's mean.}

\item{sigma}{The distribution's SD.}

\item{k}{The coefficient on the signal standard variance.}

\item{ub}{The upper truncation of the normal.}
}
\value{
The CDF at \code{x}
}
\description{
Calculate the CDF of a Normal Truncated Above
}
