% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{assessment}
\alias{assessment}
\title{Assess the output of the model}
\usage{
assessment(df_fp)
}
\arguments{
\item{df_fp}{A filepath for a excel sheet of draws}
}
\value{
A ggplot of the parameters
}
\description{
Assess the output of the model
}
