% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid_mats.R
\name{bid_mats}
\alias{bid_mats}
\title{Calculates matrices of optimal bids for interpolation.}
\usage{
bid_mats(k, n_max, x_seq, c_seq)
}
\arguments{
\item{k}{The coefficient on the signal standard variance.}

\item{n_max}{The maximum number of bidders in the auction.}

\item{x_seq}{The grid of signal values}

\item{c_seq}{The grid of default cost values}
}
\value{
A list of bidding matrices for each set of parameters.
}
\description{
Calculates matrices of optimal bids for interpolation.
}
