% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid_vec.R
\name{bid_vec}
\alias{bid_vec}
\title{Generate The Optimal Probability-Weighted Bids}
\usage{
bid_vec(mu, sigma, x_seq, c_seq, c_norm, lambda, n_max, k, mat)
}
\arguments{
\item{mu}{The auction's value of}

\item{x_i}{}
}
\description{
This function expects that all arguments are adjusted to the units of the bid matrix.
}
