% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqm.R
\name{eqm}
\alias{eqm}
\title{Equilibrium Function}
\usage{
eqm(b, mu_0, sigma_0, x_i, n, c, k)
}
\arguments{
\item{b}{The proposed bid.}

\item{mu_0}{The prior mean.}

\item{sigma_0}{The prior SD.}

\item{x_i}{The bidder's signal.}

\item{n}{The number of bidders.}

\item{c}{The cost of default.}

\item{k}{The coefficient on the signal standard variance.}
}
\value{
The error of the equilibrium function at bid \code{b}.
}
\description{
This function will take value zero for the equilibrium bid \code{b}.
}
