% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model.R
\name{estimate_model}
\alias{estimate_model}
\title{Estimate the MCMC Model}
\usage{
estimate_model(data, n_steps, start_params, n_max, par = FALSE)
}
\arguments{
\item{data}{The dataset of auction data}

\item{n_steps}{The number of MCMC draws to do}

\item{start_params}{The starting parameters for the MCMC routine.}

\item{n_max}{The maximum number of bidders.}

\item{par}{Whether or not to run the estimation in parallel}
}
\value{
The \code{draws} object from \code{\link{mcmc}}
}
\description{
Estimate the MCMC Model
}
