% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_star.R
\name{x_star}
\alias{x_star}
\title{Calculate the Threshold Signal Value}
\usage{
x_star(r, lambda, mu_0, sigma_0, k, n_max)
}
\arguments{
\item{r}{The auction's reserve price.}

\item{lambda}{The poisson parameter dictating entry.}

\item{mu_0}{The prior distribution's mean.}

\item{sigma_0}{The prior distribution's standard deviation.}

\item{k}{The parameter governing the tightness of the signal distribution.}

\item{n_max}{The maximum number of bidders in the auction.}
}
\value{
The threshold signal
}
\description{
Calculate the Threshold Signal Value
}
